// src/pages/survey/ts/enums.ts
export enum OnBoardingSteps {
  GET_STARTED = 'GET_STARTED',
  ACCOUNT_DETAILS = 'ACCOUNT_DETAILS',
  CREATE = 'CREATE',
}

export enum PatientAssessmentSteps {
  PATIENT_INFORMATION = 'PATIENT_INFORMATION',
  HEALTH_HISTORY = 'HEALTH_HISTORY',
  SOCIAL_HISTORY = 'SOCIAL_HISTORY',
  SURGICAL_HISTORY = 'SURGICAL_HISTORY',
  FAMILY_HISTORY = 'FAMILY_HISTORY',
  SYMPTOMS = 'SYMPTOMS',
}